+++ types/index.d.ts
@@ -501,6 +501,12 @@ declare namespace K {
     juxt<T = any, U = any>(fns: ReadonlyArray<(...args: T[]) => any>, args: any[]): U[];
     juxt<T = any>(fns: ReadonlyArray<(...args: T[]) => any>): <U = any>(args: any[]) => U[];
 
+
+    /**
+     * Returns a list of all the own properties of an object
+     */
+    keys<T extends object>(obj: T): Array<keyof T>;
+
     /**
      * Grabs the last index of a list
      */